src/ui/requirements.txt
Flask==3.0.3
httpx==0.27.2
python-dotenv==1.0.1